Turret created and a temporary scene
Turret class was created/finished with 99% certainty (It will need to be tweaked once we implement more functions to the game) and a bullet prefab/class was created also. An enemy class was created which was only used for testing the turret functionality and will likely be removed in the next few commits.