Fixed game breaking bug with new troops/tanks
Troop and tanks prefabs would attack and kill player 2's own buildings.

It is not a great fix but I did a little work around for just having two of each prefab, each with the playernum manually adjusted in the prefab.

We can work on changing that value during spawning later.